Book App/Book App/requirements.txt
uvicorn
fastapi
python-multipart
pymongo
python-dotenv
pydantic
pytest
httpx

